Feature

+++ src/main/java/com/project/farmeasyportal/payloads/LoanFormDTO.java
@@ -20,11 +20,11 @@ public class LoanFormDTO {
     private String district;
     private String state;
     private String pinCode;
-    private int aadhaarNumber;
+    private long aadhaarNumber;
     private String aadhaarPdfName;
-    private int panNumber;
+    private long panNumber;
     private String panPdfName;
-    private int salary;
+    private long salary;
 
     private String collateralType;
     private String guarantorName;
